requirements.txt
future       1.0.0
numpy        2.2.2
pillow       11.1.0
pip          24.2
tkintertable 1.3.3